Delete empty constructor
Hi,

We are so glad that you prepared these codes for MAX30100. It really helped us to go further on our project.

When I deleted the constructor in .cpp file, it worked fine and we got the pulse graphic clearly. I wanted to thank you and warn you about that issue.

But we couldn't do it by using RED. And also we couldn't figure out how to take SpO2 values. If you have made any progress on those topics. We would love to hear from you.

Thanks.